Solidity changes

+++ contracts/Constants.sol
@@ -11,7 +11,7 @@ library CosmicGameConstants {
 	uint256 public constant TOKEN_REWARD = 100 * 1e18;
 	uint256 public constant MARKETING_REWARD = 15 * 1e18;
 	uint256 public constant DEFAULT_AUCTION_LENGTH = 2 * 3600;
-	uint256 public constant DEFAULT_MIN_STAKE_PERIOD = 3600 * 24 * 30; // 30 days
+	uint32 public constant DEFAULT_MIN_STAKE_PERIOD = 3600 * 24 * 30; // 30 days
 	uint256 public constant MODE_RUNTIME = 0;
 	uint256 public constant MODE_PREPARE_MAINTENANCE = 1 ;
 	uint256 public constant MODE_MAINTENANCE = 2;
